Add basic plotting with uplot
At the moment the data is passed from rust to js by js allocating arrays
and giving them to rust to fill. This works but isn't ideal as js has to
know how big to make the array ahead of time. It also means that each
data array needs to be separately passed in which is not really that
nice or neat.